requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.2
django-environ==0.11.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pycparser==2.21
pytz==2024.1
setuptools==69.1.1
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.10.0
tzdata==2024.1
update==0.0.1
wheel==0.43.0
